Forum placeholder and routes
Created the placeholder for the forum and made sure it loaded dynamically on both wide and less wide displays.
Also added a header bar with intent to include the header and navigation options between the results and the forum section.
Header bar styling was also added in the style.css file.
Created the route to the forum page so it can be navigated to.